+++ packages/ui-components/web/src/Components/SettingsPanelHeader.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styles from '../../statics/styles/settings-panel-header.scss';
-import { CloseIcon } from '..';
-
+import { CloseIcon } from '../Icons';
 interface SettingsPanelHeaderProps {
   title?: string;
   onClose: () => void;
@@ -15,11 +14,13 @@ const SettingsPanelHeader: React.FC<SettingsPanelHeaderProps> = ({
   children,
   showCloseIcon = true,
 }) => {
+  const onKeyPress = e => e.key === 'Enter' && onClose();
+
   return (
     <div className={styles.settingsPanelHeader}>
       <div className={styles.settingsPanelHeader_title}>{children || title}</div>
       <div className={styles.settingsPanelHeader_closeIcon}>
-        {showCloseIcon && <CloseIcon onClick={onClose} />}
+        {showCloseIcon && <CloseIcon onClick={onClose} onKeyPress={onKeyPress} tabIndex={0} />}
       </div>
     </div>
   );